+++ packages/terra-clinical-data-grid/src/terra-dev-site/test/clinical-data-grid/NoOverflowDataGrid.test.jsx
@@ -1,6 +1,4 @@
 import React from 'react';
-
-// eslint-disable-next-line import/no-extraneous-dependencies, import/no-unresolved, import/extensions
 import DataGrid from 'terra-clinical-data-grid';
 import classNames from 'classnames/bind';
 
@@ -45,6 +43,7 @@ class NoOverflowDataGrid extends React.Component {
     };
   }
 
+  /* eslint-disable react/forbid-dom-props */
   render() {
     return (
       <div id="no-overflow-data-grid" className={cx('content-wrapper')}>
@@ -61,6 +60,7 @@ class NoOverflowDataGrid extends React.Component {
       </div>
     );
   }
+  /* eslint-enable react/forbid-dom-props */
 }
 
 export default NoOverflowDataGrid;