+++ packages/terra-clinical-data-grid/src/terra-dev-site/test/clinical-data-grid/NoPinnedColumnDataGrid.test.jsx
@@ -1,6 +1,4 @@
 import React from 'react';
-
-// eslint-disable-next-line import/no-extraneous-dependencies, import/no-unresolved, import/extensions
 import DataGrid from 'terra-clinical-data-grid';
 import classNames from 'classnames/bind';
 
@@ -86,6 +84,7 @@ class NoPinnedColumnDataGrid extends React.Component {
     };
   }
 
+  /* eslint-disable react/forbid-dom-props */
   render() {
     return (
       <div id="no-pinned-column-data-grid" className={cx('content-wrapper')}>
@@ -101,6 +100,7 @@ class NoPinnedColumnDataGrid extends React.Component {
       </div>
     );
   }
+  /* eslint-enable react/forbid-dom-props */
 }
 
 export default NoPinnedColumnDataGrid;