+++ packages/terra-clinical-data-grid/src/terra-dev-site/test/clinical-data-grid/SelectableDataGrid.test.jsx
@@ -1,6 +1,4 @@
 import React from 'react';
-
-// eslint-disable-next-line import/no-extraneous-dependencies, import/no-unresolved, import/extensions
 import DataGrid from 'terra-clinical-data-grid';
 import classNames from 'classnames/bind';
 
@@ -95,6 +93,7 @@ class SelectableDataGrid extends React.Component {
     };
   }
 
+  /* eslint-disable react/forbid-dom-props */
   render() {
     const { columns } = this.state;
 
@@ -121,12 +120,12 @@ class SelectableDataGrid extends React.Component {
           onColumnSelect={(columnId) => {
             const newColumns = {};
 
-            const columnToSort = Object.assign({}, this.state.columns[columnId]);
+            const columnToSort = { ...this.state.columns[columnId] };
             columnToSort.sortIndicator = columnToSort.sortIndicator === 'ascending' ? 'descending' : 'ascending';
             newColumns[`${columnId}`] = columnToSort;
 
             if (columnId !== this.state.sortedColumnId) {
-              const previouslySortedColumn = Object.assign({}, this.state.columns[this.state.sortedColumnId]);
+              const previouslySortedColumn = { ...this.state.columns[this.state.sortedColumnId] };
               if (previouslySortedColumn) {
                 previouslySortedColumn.sortIndicator = undefined;
                 newColumns[`${this.state.sortedColumnId}`] = previouslySortedColumn;
@@ -134,7 +133,7 @@ class SelectableDataGrid extends React.Component {
             }
 
             this.setState(prevState => ({
-              columns: Object.assign({}, prevState.columns, newColumns),
+              columns: { ...prevState.columns, ...newColumns },
               sortedColumnId: columnId,
               sortedColumnDirection: columnToSort.sortIndicator,
             }));
@@ -163,6 +162,7 @@ class SelectableDataGrid extends React.Component {
       </div>
     );
   }
+  /* eslint-enable react/forbid-dom-props */
 }
 
 export default SelectableDataGrid;