+++ packages/terra-clinical-data-grid/src/terra-dev-site/test/clinical-data-grid/StandardDataGrid.test.jsx
@@ -1,6 +1,4 @@
 import React from 'react';
-
-// eslint-disable-next-line import/no-extraneous-dependencies, import/no-unresolved, import/extensions
 import DataGrid from 'terra-clinical-data-grid';
 import classNames from 'classnames/bind';
 
@@ -89,6 +87,7 @@ class StandardDataGrid extends React.Component {
     };
   }
 
+  /* eslint-disable react/forbid-dom-props */
   render() {
     return (
       <div id="standard-data-grid" className={cx('content-wrapper')}>
@@ -105,6 +104,7 @@ class StandardDataGrid extends React.Component {
       </div>
     );
   }
+  /* eslint-enable react/forbid-dom-props */
 }
 
 export default StandardDataGrid;