+++ packages/terra-clinical-detail-view/src/terra-dev-site/test/clinical-detail-view/DetailView/DetailViewNoDivider.test.jsx
@@ -17,7 +17,7 @@ const DetailViewNoDivider = () => (
     subtitles={['Subtitle 1', 'Subtitle 2']}
     graph={(
       <div className={cx('graph-content')}>
-      This is where any visualizations would go
+        This is where any visualizations would go
       </div>
 )}
     accessory={accessoryElement}