+++ packages/terra-clinical-header/src/terra-dev-site/doc/clinical-header/clinicalHeader.1.doc.jsx
@@ -3,7 +3,6 @@ import DocTemplate from 'terra-doc-template';
 import ReadMe from '../../../../docs/README.md';
 import { name } from '../../../../package.json';
 
-/* eslint-disable import/no-webpack-loader-syntax, import/first, import/extensions, import/no-unresolved, import/no-duplicates */
 // Component Source
 import HeaderSrc from '!raw-loader!../../../../src/Header';
 
@@ -18,7 +17,6 @@ import HeaderLongTextWithContent from '../example/HeaderLongTextWithContent';
 import HeaderLongTextWithContentSrc from '!raw-loader!../../../../src/terra-dev-site/doc/example/HeaderLongTextWithContent';
 import Subheader from '../example/Subheader';
 import SubheaderSrc from '!raw-loader!../../../../src/terra-dev-site/doc/example/Subheader';
-/* eslint-enable import/no-webpack-loader-syntax, import/first, import/extensions, import/no-unresolved, import/no-duplicates */
 
 const DocPage = () => (
   <DocTemplate