+++ packages/terra-clinical-item-view/src/terra-dev-site/test/clinical-item-view/DefaultItemView.test.jsx
@@ -5,6 +5,7 @@ import styles from './ItemViewCommon.test.module.scss';
 
 const cx = classNames.bind(styles);
 
+/* eslint-disable react/forbid-dom-props */
 const DefaultItemView = () => (
   <div>
     <p>Given no props, the div is created, but there is no content to show. The ItemView is outlined below.</p>
@@ -13,5 +14,6 @@ const DefaultItemView = () => (
     </div>
   </div>
 );
+/* eslint-enable react/forbid-dom-props */
 
 export default DefaultItemView;