+++ packages/terra-clinical-label-value-view/src/terra-dev-site/test/clinical-label-value-view/TextValueLabelValueView.test.jsx
@@ -6,6 +6,7 @@ import styles from './TextValueLabelValueView.test.module.scss';
 
 const cx = classNames.bind(styles);
 
+/* eslint-disable react/forbid-dom-props */
 const examples = () => (
   <div>
     <h3>Default Example:</h3>
@@ -16,5 +17,6 @@ const examples = () => (
     </div>
   </div>
 );
+/* eslint-enable react/forbid-dom-props */
 
 export default examples;